+++ src/store/slices/index.ts
@@ -1,2 +1,2 @@
 // re-name and export the slices
-export * as exampleThunk from './example-thunk';
+export * as exampleThunk from './exampleThunkSlice';
